refactor: Secondary Index를 통한 Redis 저장 구조 개편
