+++ src/main/java/com/waruru/areyouhere/active/ActiveSessionAttendanceController.java
@@ -1,9 +1,9 @@
 package com.waruru.areyouhere.active;
 
 
-import com.waruru.areyouhere.attendance.service.AttendanceService;
 import com.waruru.areyouhere.attendance.dto.request.AuthCodeDeactivationRequestDto;
 import com.waruru.areyouhere.attendance.dto.request.AuthCodeRequestDto;
+import com.waruru.areyouhere.active.service.ActiveSessionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,19 +14,19 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(ActiveSessionController.AUTH_CODE_API_URL)
-public class ActiveSessionController {
+@RequestMapping(ActiveSessionAttendanceController.AUTH_CODE_API_URL)
+public class ActiveSessionAttendanceController {
 
     public static final String AUTH_CODE_API_URL = "/api/auth-code";
 
-    private final AttendanceService attendanceService;
+    private final ActiveSessionService activeSessionService;
 
 
     @PostMapping
     public ResponseEntity<String> activate(@RequestBody AuthCodeRequestDto authCodeRequestDto) {
         Long sessionId = authCodeRequestDto.getSessionId();
         Long courseId = authCodeRequestDto.getCourseId();
-        return ResponseEntity.ok(attendanceService.activateSession(sessionId, courseId));
+        return ResponseEntity.ok(activeSessionService.activateSession(sessionId, courseId));
     }
 
     @PostMapping("/deactivate")
@@ -35,7 +35,7 @@ public class ActiveSessionController {
         Long sessionId = authCodeDeactivationRequestDto.getSessionId();
         Long courseId = authCodeDeactivationRequestDto.getCourseId();
         String authCode = authCodeDeactivationRequestDto.getAuthCode();
-        attendanceService.deactivateSession(authCode, sessionId, courseId);
+        activeSessionService.deactivateSession(authCode, sessionId, courseId);
         return ResponseEntity.ok().build();
     }
