--- src/main/java/com/waruru/areyouhere/active/domain/repository/AttendanceRedisRepository.java
@@ -1,36 +1,0 @@
-package com.waruru.areyouhere.active.domain.repository;
-
-import com.waruru.areyouhere.attendance.dto.AttendeeRedisData;
-import java.util.Set;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class AttendanceRedisRepository {
-    private final RedisTemplate<String, Long> redisTemplate;
-
-    private final String ATTENDANCE_KEY = "attendance";
-
-    public boolean isAlreadyAttended(String authCode, AttendeeRedisData attendeeInfo){
-        Set<Long> attendedMembers = redisTemplate.opsForSet().members(ATTENDANCE_KEY + authCode);
-        return attendedMembers == null || attendedMembers.contains(attendeeInfo.getId());
-    }
-
-    public Set<Long> getAttendees(String authCode){
-        return redisTemplate.opsForSet().members(ATTENDANCE_KEY + authCode);
-    }
-
-    public void setAttend(String authCode, AttendeeRedisData attendeeInfo){
-        redisTemplate.opsForSet().add(ATTENDANCE_KEY + authCode, attendeeInfo.getId());
-    }
-
-    public void deleteAllAttendanceInSession(String authCode){
-        redisTemplate.delete(ATTENDANCE_KEY + authCode);
-    }
-
-
-}
-
-
