--- src/main/java/com/waruru/areyouhere/active/domain/repository/CourseIdRedisRepository.java
@@ -1,7 +1,0 @@
-package com.waruru.areyouhere.active.domain.repository;
-
-import com.waruru.areyouhere.active.domain.entity.CourseId;
-import org.springframework.data.repository.CrudRepository;
-
-public interface CourseIdRedisRepository extends CrudRepository<CourseId, Long> {
-}
