--- src/main/java/com/waruru/areyouhere/active/domain/repository/SessionIdRedisRepository.java
@@ -1,9 +1,0 @@
-package com.waruru.areyouhere.active.domain.repository;
-
-import com.waruru.areyouhere.active.domain.entity.SessionId;
-import org.springframework.data.repository.CrudRepository;
-
-public interface SessionIdRedisRepository extends CrudRepository<SessionId, Long> {
-
-
-}
