+++ src/main/java/com/waruru/areyouhere/attendance/service/AttendanceService.java
@@ -15,8 +15,4 @@ public interface AttendanceService {
     public CurrentSessionAttendCount getCurrentSessionAttendCount(Long sessionId);
 
     public CurrentSessionAttendeeAttendance getCurrentSessionAttendeesAndAbsentees(String authCode);
-
-    public String activateSession(Long sessionId, Long courseId);
-
-    public void deactivateSession(String authCode, Long sessionId, Long courseId);
 }
