+++ src/main/java/com/waruru/areyouhere/attendance/service/AttendanceServiceImpl.java
@@ -6,14 +6,9 @@ import com.waruru.areyouhere.attendance.dto.response.AttendResponseDto;
 import com.waruru.areyouhere.attendance.service.dto.CurrentSessionAttendCount;
 import com.waruru.areyouhere.attendance.service.dto.CurrentSessionAttendeeAttendance;
 import com.waruru.areyouhere.attendance.service.rdb.AttendanceRDBService;
-import com.waruru.areyouhere.active.service.ActiveSessionService;
+import com.waruru.areyouhere.active.service.ActiveAttendanceService;
 import com.waruru.areyouhere.attendee.service.dto.AttendeeInfo;
-import com.waruru.areyouhere.course.domain.entity.Course;
-import com.waruru.areyouhere.course.service.CourseService;
-import com.waruru.areyouhere.session.domain.entity.Session;
-import com.waruru.areyouhere.session.service.command.SessionCommandService;
 import com.waruru.areyouhere.session.service.dto.AuthCodeInfo;
-import com.waruru.areyouhere.session.service.query.SessionQueryService;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.List;
@@ -27,16 +22,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AttendanceServiceImpl implements AttendanceService {
     private final AttendanceRDBService attendanceRDBService;
-    private final ActiveSessionService activeSessionService;
-    private final CourseService courseService;
-    private final SessionQueryService sessionQueryService;
-    private final SessionCommandService sessionCommandService;
+    private final ActiveAttendanceService activeAttendanceService;
 
     @Override
     @Transactional
     public AttendResponseDto attend(String attendeeName, String authCode, Long attendeeId) {
         LocalDateTime attendanceTime = LocalDateTime.now(ZoneId.of("Asia/Seoul"));
-        List<AttendeeInfo> nameSakeAttendees = activeSessionService.getNameSakeInfos(authCode, attendeeName);
+        List<AttendeeInfo> nameSakeAttendees = activeAttendanceService.getNameSakeInfos(authCode, attendeeName);
 
         // 동명이인 응답
         if (attendeeId == null && nameSakeAttendees.size() > 1) {
@@ -45,11 +37,11 @@ public class AttendanceServiceImpl implements AttendanceService {
                     .build();
         }
 
-        AuthCodeInfo authCodeInfo = activeSessionService.isAttendPossible(authCode, attendeeName, attendeeId);
-        AttendeeRedisData attendeeInSession = activeSessionService.findByNameIfNotDuplicatedOrId(attendeeName,
+        AuthCodeInfo authCodeInfo = activeAttendanceService.isAttendPossible(authCode, attendeeName, attendeeId);
+        AttendeeRedisData attendeeInSession = activeAttendanceService.findByNameIfNotDuplicatedOrId(attendeeName,
                 attendeeId,
-                activeSessionService.getSessionAttendanceInfoOrThrow(authCode));
-        activeSessionService.setAttendInRedis(authCode, attendeeInSession);
+                activeAttendanceService.getSessionAttendanceInfoOrThrow(authCode));
+        activeAttendanceService.setAttendInRedis(authCode, attendeeInSession);
 
         return AttendResponseDto.builder()
                 .attendanceName(attendeeName)
@@ -68,38 +60,18 @@ public class AttendanceServiceImpl implements AttendanceService {
     @Override
     @Transactional(readOnly = true)
     public CurrentSessionAttendCount getCurrentSessionAttendCount(Long sessionId) {
-        String authCode = activeSessionService.findAuthCodeBySessionId(sessionId);
-        int total = activeSessionService.getTotalAttendees(authCode);
-        int attendeeCount = activeSessionService.getAttendCount(authCode);
+        String authCode = activeAttendanceService.findAuthCodeBySessionId(sessionId);
+        int total = activeAttendanceService.getTotalAttendees(authCode);
+        int attendeeCount = activeAttendanceService.getAttendCount(authCode);
         return new CurrentSessionAttendCount(total, attendeeCount);
     }
 
     @Override
     @Transactional(readOnly = true)
     public CurrentSessionAttendeeAttendance getCurrentSessionAttendeesAndAbsentees(String authCode) {
-        return activeSessionService.getCurrentSessionAttendees(authCode);
+        return activeAttendanceService.getCurrentSessionAttendees(authCode);
     }
 
-    @Override
-    @Transactional
-    public String activateSession(Long sessionId, Long courseId) {
-        Course course = courseService.get(courseId);
-        Session session = sessionQueryService.get(sessionId);
-        LocalDateTime currentTime = LocalDateTime.now(ZoneId.of("Asia/Seoul"));
-        return activeSessionService.activate(course, session, currentTime);
-    }
-
-    @Override
-    @Transactional
-    public void deactivateSession(String authCode, Long sessionId, Long courseId) {
-        sessionQueryService.checkNotDeactivated(sessionId);
-        CurrentSessionAttendeeAttendance currentSessionAttendees = activeSessionService.getCurrentSessionAttendees(
-                authCode);
-
-        attendanceRDBService.setAttendancesAfterDeactivate(courseId, sessionId, currentSessionAttendees);
-        sessionCommandService.deactivate(sessionId);
-        activeSessionService.deactivate(authCode);
-    }
 
 
 }
