+++ src/main/java/com/waruru/areyouhere/attendance/service/rdb/AttendanceRDBServiceImpl.java
@@ -1,6 +1,7 @@
 package com.waruru.areyouhere.attendance.service.rdb;
 
 
+import com.waruru.areyouhere.active.domain.entity.CurrentSessionAttendanceInfo;
 import com.waruru.areyouhere.attendance.domain.entity.Attendance;
 import com.waruru.areyouhere.attendance.domain.repository.AttendanceBatchRepository;
 import com.waruru.areyouhere.attendance.domain.repository.AttendanceRepository;
@@ -12,7 +13,11 @@ import com.waruru.areyouhere.attendee.domain.repository.AttendeeRepository;
 import com.waruru.areyouhere.session.domain.entity.Session;
 import com.waruru.areyouhere.session.domain.repository.SessionRepository;
 import com.waruru.areyouhere.session.exception.SessionIdNotFoundException;
+import java.time.LocalDateTime;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,13 +36,25 @@ public class AttendanceRDBServiceImpl implements AttendanceRDBService {
 
 
     @Override
-    public void setAttendancesAfterDeactivate(long courseId, long sessionId, CurrentSessionAttendeeAttendance currentSessionAttendeeAttendance) {
+    public void setAttendancesAfterDeactivate(long courseId, long sessionId,
+                                              CurrentSessionAttendanceInfo currentSessionAttendanceInfo) {
 
-        List<AttendeeRedisData> attendees = currentSessionAttendeeAttendance.getAttendees();
-        List<AttendeeRedisData> absentees = currentSessionAttendeeAttendance.getAbsentees();
+        LocalDateTime absentTime = LocalDateTime.now();
+        List<AttendeeRedisData> attendees = new LinkedList<>();
+        List<AttendeeRedisData> absentees = new LinkedList<>();
+        Map<Long, LocalDateTime> attendanceTime = currentSessionAttendanceInfo.getAttendanceTime();
+        currentSessionAttendanceInfo.getAttendees()
+                .forEach(att -> {
+                    if (attendanceTime.containsKey(att.getId())) {
+                        attendees.add(att);
+                    } else {
+                        absentees.add(att);
+                    }
+                });
 
-        attendanceBatchRepository.insertAttendanceBatch(attendees, true, sessionId);
-        attendanceBatchRepository.insertAttendanceBatch(absentees, false, sessionId);
+
+        attendanceBatchRepository.insertAttendBatch(attendees, true, sessionId, attendanceTime);
+        attendanceBatchRepository.insertAbsentBatch(absentees, false, sessionId, absentTime);
     }
 
 
@@ -53,30 +70,6 @@ public class AttendanceRDBServiceImpl implements AttendanceRDBService {
                 .collect(Collectors.toList());
 
         attendanceRepository.saveAll(attendancesToUpdate);
-    }
-
-    @Transactional(readOnly = true)
-    public int currentAttendance(Long sessionId) {
-        List<Attendance> attendancesBySessionId = attendanceRepository.findAttendancesBySession_Id(sessionId);
-        if (attendancesBySessionId == null || attendancesBySessionId.isEmpty()) {
-            return 0;
-        } else {
-            return attendancesBySessionId.size();
-        }
-    }
-
-
-    // AttendeeId가 null이 아니라는 것은 duplicatedAttendee가 발생하여 Id 기준으로 찾아야 한다는 것.
-    private Attendee getAttendee(String attendanceName, Long attendeeId, List<Attendee> attendeesByCourseId) {
-        return attendeeId == null ?
-                attendeesByCourseId.stream()
-                        .filter(s -> s.getName().equals(attendanceName))
-                        .findFirst()
-                        .orElse(null) :
-                attendeesByCourseId.stream()
-                        .filter(s -> s.getId().equals(attendeeId))
-                        .findFirst()
-                        .orElse(null);
     }
